move initialization from sim to pars
