+++ starsim/sim.py
@@ -6,7 +6,6 @@ Define core Sim classes
 import numpy as np
 import sciris as sc
 import starsim as ss
-import itertools
 import matplotlib.pyplot as pl
 
 __all__ = ['Sim', 'AlreadyRunError', 'demo', 'diff_sims']
@@ -15,98 +14,43 @@ __all__ = ['Sim', 'AlreadyRunError', 'demo', 'diff_sims']
 class Sim(sc.prettyobj):
 
     def __init__(self, pars=None, label=None, people=None, demographics=None, diseases=None, networks=None,
-                 connectors=None, interventions=None, analyzers=None, **kwargs):
+                 interventions=None, analyzers=None, connectors=None, copy_inputs=True, **kwargs):
 
-        # Set attributes
-        self.label = label  # The label/name of the simulation
-        self.created = None  # The datetime the sim was created
-        self.people = people  # People object
+        # Make default parameters (using values from parameters.py)
+        self.pars = ss.make_pars() # Start with default pars
+        args = dict(label=label, people=people, demographics=demographics, diseases=diseases, networks=networks, 
+                    interventions=interventions, analyzers=analyzers, connectors=connectors)
+        args = {key:val for key,val in args.items() if val is not None} # Remove None inputs
+        self.pars.update_pars(sc.mergedicts(args, pars, kwargs, _copy=copy_inputs))  # Update the parameters
         
-        self.summary = None  # For storing a summary of the results
+        # Set attributes
+        self.created = sc.now()  # The datetime the sim was created
         self.initialized = False  # Whether initialization is complete
         self.complete = False  # Whether a simulation has completed running # TODO: replace with finalized?
         self.results_ready = False  # Whether results are ready
-        self.filename = None
-
-        # Time indexing
-        self.ti = None  # The time index, e.g. 0, 1, 2
-        self.yearvec = None
-        self.tivec = None
-        self.npts = None
-
-        # Make default parameters (using values from parameters.py)
-        self.pars = ss.make_pars()  # Start with default pars
-        self.pars.update_pars(sc.mergedicts(pars, kwargs))  # Update the parameters
-
-        # Placeholders for plug-ins: demographics, diseases, connectors, analyzers, and interventions
-        # Products are not here because they are stored within interventions
-        if demographics == True:
-            demographics = [ss.Births(), ss.Deaths()]  # Use default assumptions for demographics
-        self.demographics  = ss.ndict(demographics, type=ss.Demographics)
-        self.diseases      = ss.ndict(diseases, type=ss.Disease)
-        self.networks      = ss.ndict(networks, type=ss.Network)
-        self.connectors    = ss.ndict(connectors, type=ss.Connector)
-        self.interventions = ss.ndict(interventions, type=ss.Intervention, strict=False) # strict=False since can be a function
-        self.analyzers     = ss.ndict(analyzers, type=ss.Analyzer, strict=False)
-
-        # Initialize the random number generator container
-        self.dists = ss.Dists(obj=self)
+        self.dists = ss.Dists(obj=self) # Initialize the random number generator container
         self.results = ss.Results(module='sim')  # For storing results
-
+        self.summary = None  # For storing a summary of the results
+        self.filename = None # Store the filename, if saved
         return
     
-    @property
-    def label(self):
-        return self.pars.get('label', '')
-
-    @property
-    def dt(self):
-        return self.pars.get('dt', np.Nan)
-
-    @property
-    def year(self):
-        try:
-            return self.yearvec[self.ti]
-        except:
-            return np.nan
-
-    @property
-    def modules(self):
-        # Return iterator over all Module instances (stored in standard places) in the Sim
-        products = [intv.product for intv in self.interventions.values() if
-                    hasattr(intv, 'product') and isinstance(intv.product, ss.Product)]
-        return itertools.chain(
-            self.demographics.values(),
-            self.networks.values(),
-            self.diseases.values(),
-            self.connectors.values(),
-            self.interventions.values(),
-            products,
-            self.analyzers.values(),
-        )
-
     def initialize(self, reset=False, **kwargs):
-        """
-        Perform all initializations on the sim.
-        """
+        """ Perform all initializations for the sim; most heavy lifting is done by the parameters """
         # Validation and initialization
-        self.ti = 0  # The current time index
-        self.validate_pars()  # Ensure parameters have valid values
-        self.validate_dt()
-        self.init_time_vecs()  # Initialize time vecs
-        ss.set_seed(self.pars.rand_seed)  # Reset the seed before the population is created -- shouldn't matter if only using Dist objects
-
-        # Initialize the core sim components
-        self.init_people(reset=reset, **kwargs)  # Create all the people (the heaviest step)
-
-        # Initialize plug-ins
-        self.init_demographics()
-        self.init_networks()
-        self.init_diseases()
-        self.init_connectors()
-        self.init_interventions()
-        self.init_analyzers()
-
+        ss.set_seed(self.pars.rand_seed) # Reset the seed before the population is created -- shouldn't matter if only using Dist objects
+        self.pars.initialize(sim=self, reset=reset, **kwargs) # Initialize the parameters, including people and modules
+        
+        # Move initialized modules to the sim
+        keys = ['dt', 'label', 'people', 'demographics', 'networks', 'diseases', 'interventions', 'analyzers', 'connectors']
+        for key in keys:
+            setattr(self, key, self.pars.pop(key))
+            
+        # Time indexing
+        self.yearvec = np.arange(start=self.pars.start, stop=self.pars.end, step=self.dt)
+        self.npts = len(self.yearvec)
+        self.tivec = np.arange(self.npts)
+        self.ti = 0  # The time index, e.g. 0, 1, 2
+        
         # Initialize all distributions now that everything else is in place
         self.dists.initialize(obj=self, base_seed=self.pars.rand_seed, force=True)
 
@@ -116,349 +60,6 @@ class Sim(sc.prettyobj):
         self.results_ready = False
 
         return self
-
-    def validate_dt(self):
-        """
-        Check that 1/dt is an integer value, otherwise results and time vectors will have mismatching shapes.
-        init_results explicitly makes this assumption by casting resfrequency = int(1/dt).
-        """
-        dt = self.dt
-        reciprocal = 1.0 / dt  # Compute the reciprocal of dt
-        if not reciprocal.is_integer():  # Check if reciprocal is not a whole (integer) number
-            # Round the reciprocal
-            reciprocal = int(reciprocal)
-            rounded_dt = 1.0 / reciprocal
-            self.pars.dt = rounded_dt
-            if self.pars.verbose:
-                warnmsg = f"Warning: Provided time step dt: {dt} resulted in a non-integer number of steps per year. Rounded to {rounded_dt}."
-                print(warnmsg)
-        return
-
-    def validate_pars(self):
-        """
-        Some parameters can take multiple types; this makes them consistent.
-        """
-        # Handle n_agents
-        if self.people is not None:
-            self.pars.n_agents = len(self.people)
-        # elif self.popdict is not None: # Starsim does not currenlty support self.popdict
-        # self.pars.n_agents = len(self.popdict)
-        elif self.pars.n_agents is not None:
-            self.pars.n_agents = int(self.pars.n_agents)
-        else:
-            errormsg = 'Must supply n_agents, a people object, or a popdict'
-            raise ValueError(errormsg)
-
-        # Handle end and n_years
-        if self.pars.end:
-            self.pars.n_years = self.pars.end - self.pars.start
-            if self.pars.n_years <= 0:
-                errormsg = f"Number of years must be >0, but you supplied start={str(self.pars.start)} and " \
-                           f"end={str(self.pars.end)}, which gives n_years={self.pars.n_years}"
-                raise ValueError(errormsg)
-        else:
-            if self.pars.n_years:
-                self.pars.end = self.pars.start + self.pars.n_years
-            else:
-                errormsg = 'You must supply one of n_years and end."'
-                raise ValueError(errormsg)
-
-        # Handle verbose
-        if self.pars.verbose == 'brief':
-            self.pars.verbose = -1
-        if not sc.isnumber(self.pars.verbose):  # pragma: no cover
-            errormsg = f'Verbose argument should be either "brief", -1, or a float, not {type(self.par.verbose)} "{self.par.verbose}"'
-            raise ValueError(errormsg)
-
-        return
-
-    def init_time_vecs(self):
-        """
-        Construct vectors things that keep track of time
-        """
-        self.yearvec = np.arange(start=self.pars.start, stop=self.pars.end + self.pars.dt, step=self.pars.dt)
-        self.npts = len(self.yearvec)
-        self.tivec = np.arange(self.npts)
-        return
-
-    def init_people(self, reset=False, verbose=None, **kwargs):
-        """
-        Initialize people within the sim
-        Sometimes the people are provided, in which case this just adds a few sim properties to them.
-        Other time people are not provided and this method makes them.
-        Args:
-            reset           (bool): whether to regenerate the people even if they already exist
-            verbose         (int):  detail to print
-            kwargs          (dict): passed to ss.make_people()
-        """
-
-        # Handle inputs
-        if verbose is None:
-            verbose = self.pars.verbose
-        if verbose > 0:
-            resetstr = ''
-            if self.people and reset:
-                resetstr = ' (resetting people)'
-            print(f'Initializing sim{resetstr} with {self.pars["n_agents"]:0n} agents')
-
-        # If people have not been supplied, make them
-        if self.people is None or reset:
-            self.people = ss.People(n_agents=self.pars.n_agents, **kwargs)  # This just assigns UIDs and length
-
-        # If a popdict has not been supplied, we can make one from location data
-        if self.pars.location is not None:
-            # Check where to get total_pop from
-            if self.pars.total_pop is not None:  # If no pop_scale has been provided, try to get it from the location
-                errormsg = 'You can either define total_pop explicitly or via the location, but not both'
-                raise ValueError(errormsg)
-
-        else:
-            if self.pars.total_pop is not None:  # If no pop_scale has been provided, try to get it from the location
-                total_pop = self.pars.total_pop
-            else:
-                if self.pars.pop_scale is not None:
-                    total_pop = self.pars.pop_scale * self.pars.n_agents
-                else:
-                    total_pop = self.pars.n_agents
-
-        self.pars.total_pop = total_pop
-        if self.pars.pop_scale is None:
-            self.pars.pop_scale = total_pop / self.pars.n_agents
-
-        # Any other initialization
-        if not self.people.initialized:
-            self.people.initialize(self)
-
-        # Set time attributes
-        self.people.ti = self.ti
-        self.people.dt = self.dt
-        self.people.year = self.year
-        self.people.init_results(self)
-        return self
-
-    def convert_plugins(self, plugin_class, plugin_name=None):
-        """
-        Common logic for converting plug-ins to a standard format
-        Used for networks, demographics, diseases, connectors, analyzers, and interventions
-        Args:
-            plugin: class
-        """
-
-        if plugin_name is None: plugin_name = plugin_class.__name__.lower()
-
-        # Get lower-case names of all subclasses
-        known_plugins = {n.__name__.lower():n for n in ss.all_subclasses(plugin_class)}
-        if plugin_name == 'networks': # Allow "msm" or "msmnet"
-            known_plugins.update({k.removesuffix('net'):v for k,v in known_plugins.items()})
-
-        # Figure out if it's in the sim pars or provided directly
-        attr_plugins = getattr(self, plugin_name)  # Get any plugins that have been provided directly
-
-        # See if they've been provided in the pars dict
-        if self.pars.get(plugin_name):
-
-            par_plug = self.pars[plugin_name]
-
-            # String: convert to ndict
-            if isinstance(par_plug, str):
-                plugins = ss.ndict(dict(name=par_plug))
-
-            # List or dict: convert to ndict
-            elif sc.isiterable(par_plug) and len(par_plug):
-                if isinstance(par_plug, dict) and 'type' in par_plug and 'name' not in par_plug:
-                    par_plug['name'] = par_plug['type'] # TODO: simplify/remove this
-                plugins = ss.ndict(par_plug)
-
-        else:  # Not provided directly or in pars
-            plugins = {}
-
-        # Check that we don't have two copies
-        for attr_key in attr_plugins.keys():
-            if plugins.get(attr_key):
-                errormsg = f'Sim was created with {attr_key} module, cannot create another through the pars dict.'
-                raise ValueError(errormsg)
-
-        plugins = sc.mergedicts(plugins, attr_plugins)
-
-        # Process
-        processed_plugins = sc.autolist()
-        for plugin in plugins.values():
-
-            if not isinstance(plugin, plugin_class):
-
-                if isinstance(plugin, dict):
-                    ptype = (plugin.get('type') or plugin.get('name') or '').lower()
-                    name = plugin.get('name') or ptype
-                    if ptype in known_plugins:
-                        # Make an instance of the requested plugin
-                        plugin_pars = {k: v for k, v in plugin.items() if k not in ['type', 'name']}
-                        pclass = known_plugins[ptype]
-                        plugin = pclass(name=name, pars=plugin_pars) # TODO: does this handle par_dists, etc?
-                    else:
-                        errormsg = (f'Could not convert {plugin} to an instance of class {plugin_name}.'
-                                    f'Try specifying it directly rather than as a dictionary.')
-                        raise ValueError(errormsg)
-                elif plugin_name in ['analyzers', 'interventions'] and callable(plugin):
-                    pass # This is ok, it's a function instead of an Intervention object
-                else:
-                    errormsg = (
-                        f'{plugin_name.capitalize()} must be provided as either class instances or dictionaries with a '
-                        f'"name" key corresponding to one of these known subclasses: {known_plugins}.')
-                    raise ValueError(errormsg)
-
-            processed_plugins += plugin
-
-        return processed_plugins
-
-    def init_demographics(self):
-        """ Initialize demographics """
-
-        # Demographics can be provided via sim.demographics or sim.pars - this methods reconciles them
-        demographics = self.convert_plugins(ss.Demographics, plugin_name='demographics')
-
-        # We also allow users to add vital dynamics by entering birth_rate and death_rate parameters directly to the sim
-        if self.pars.birth_rate is not None:
-            births = ss.Births(pars={'birth_rate': self.pars.birth_rate})
-            demographics += births
-        if self.pars.death_rate is not None:
-            background_deaths = ss.Deaths(pars={'death_rate': self.pars.death_rate})
-            demographics += background_deaths
-
-        # Iterate over demographic modules and initialize them
-        for dem_mod in demographics:
-            dem_mod.initialize(self)
-            self.results[dem_mod.name] = dem_mod.results
-
-        # Count how many of each kind of demographic module we have
-        demdict = {'births': ss.Births, 'pregnancy': ss.Pregnancy, 'deaths': ss.Deaths}
-        mod_names = dict()
-        for demname, demtype in demdict.items():
-            mod_names[demname] = [d.name for d in demographics if isinstance(d, demtype)]
-
-            # Validation
-            if len(mod_names[demname]) > 1:
-                if len(mod_names[demname]) == len(set(mod_names[demname])):  # No duplicate names, raise warning
-                    ss.warn(f'Two instances of {demname} module added to the sim; was this intentional?')
-                else:
-                    errormsg = (f'Cannot add two identically-named {demname} modules to a sim.\n '
-                                f'Demographic modules are: \n{sc.newlinejoin(mod_names[demname])}.\n'
-                                f'Tip: if using demographic modules, do not use birth and death rates in the sim pars.')
-                    raise ValueError(errormsg)
-
-        # Ensure they're stored at the sim level
-        self.demographics = ss.ndict(*demographics)
-
-    def init_diseases(self):
-        """ Initialize diseases """
-
-        # Diseases can be provided in sim.demographics or sim.pars
-        diseases = self.convert_plugins(ss.Disease, plugin_name='diseases')
-
-        # Interate over diseases and initialize them
-        for disease in diseases:
-            disease.initialize(self)
-
-            # Add the disease's parameters and results into the Sim's dicts
-            self.pars[disease.name] = disease.pars
-            self.results[disease.name] = disease.results
-
-            # Add disease states to the People's dicts
-            self.people.add_module(disease)
-
-        # Store diseases in the sim
-        self.diseases = ss.ndict(*diseases)
-
-        return
-
-    def init_connectors(self):
-        for connector in self.connectors.values():
-            connector.initialize(self)
-        return
-
-    def init_networks(self):
-        """ Initialize networks if these have been provided separately from the people """
-
-        processed_networks = self.convert_plugins(ss.Network, plugin_name='networks')
-
-        # Now store the networks in a Networks object, which also allows for connectors between networks
-        if not isinstance(processed_networks, ss.Networks):
-            self.networks = ss.Networks(*processed_networks)
-        self.networks.initialize(self)
-
-        return
-
-    def init_interventions(self):
-        """ Initialize and validate the interventions """
-
-        interventions = self.convert_plugins(ss.Intervention, plugin_name='interventions')
-
-        # Translate the intervention specs into actual interventions
-        for i, intervention in enumerate(interventions):
-            if isinstance(intervention, type) and issubclass(intervention, ss.Intervention):
-                intervention = intervention()  # Convert from a class to an instance of a class
-            if isinstance(intervention, ss.Intervention):
-                intervention.initialize(self)
-            elif callable(intervention):
-                intv_func = intervention
-                intervention = ss.Intervention(name=f'intervention_func_{i}')
-                intervention.apply = intv_func # Monkey-patch together an intervention from a function
-            else:
-                errormsg = f'Intervention {intervention} does not seem to be a valid intervention: must be a function or Intervention subclass'
-                raise TypeError(errormsg)
-            
-            if intervention.name not in self.interventions:
-                self.interventions += intervention
-
-            # Add the intervention parameters and results into the Sim's dicts
-            self.pars[intervention.name] = intervention.pars
-            self.results[intervention.name] = intervention.results
-
-            # Add intervention states to the People's dicts
-            self.people.add_module(intervention)
-
-            # If there's a product module present, initialize and add it
-            if hasattr(intervention, 'product') and isinstance(intervention.product, ss.Product):
-                intervention.product.initialize(self)
-
-                self.people.add_module(intervention.product)
-        
-        # TODO: combine this with the code above
-        for k,intervention in self.interventions.items():
-            if not isinstance(intervention, ss.Intervention):
-                intv_func = intervention
-                intervention = ss.Intervention(name=f'intervention_func_{k}')
-                intervention.apply = intv_func # Monkey-patch together an intervention from a function
-                self.interventions[k] = intervention
-
-        return
-
-    def init_analyzers(self):
-        """ Initialize the analyzers """
-        
-        analyzers = self.pars.analyzers
-        if not np.iterable(analyzers):
-            analyzers = sc.tolist(analyzers)
-
-        # Interpret analyzers
-        for ai, analyzer in enumerate(analyzers):
-            if isinstance(analyzer, type) and issubclass(analyzer, ss.Analyzer):
-                analyzer = analyzer()  # Convert from a class to an instance of a class
-            if not (isinstance(analyzer, ss.Analyzer) or callable(analyzer)):
-                errormsg = f'Analyzer {analyzer} does not seem to be a valid analyzer: must be a function or Analyzer subclass'
-                raise TypeError(errormsg)
-            self.analyzers += analyzer  # Add it in
-
-        # TODO: should tidy/remove this code
-        for k,analyzer in self.analyzers.items():
-            if not isinstance(analyzer, ss.Analyzer) and callable(analyzer):
-                ana_func = analyzer
-                analyzer = ss.Analyzer(name=f'analyzer_func_{k}')
-                analyzer.apply = ana_func # Monkey-patch together an intervention from a function
-                self.analyzers[k] = analyzer
-            if isinstance(analyzer, ss.Analyzer):
-                analyzer.initialize(self)
-
-        return
 
     def step(self):
         """ Step through time and update values """
